+++ CHANGELOG.md
@@ -12,6 +12,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [1.0.2] - 2025-04-03
+
+### Added
+
+- Examples Directory
+
+### Changed
+
+- "Query" param added to Get-AxoniusQueries to search for Axonius Queries using AQL, and fixed folder param getting added twice
+
 ## [1.0.1] - 2025-04-01
 
 ### Changed